requirements.txt
# 图像颜色空间分析器依赖包

# 图像处理
Pillow>=10.0.0
imageio>=2.31.0  # 用于读取TIFF等多种图像格式

# 数值计算
numpy>=1.24.0

# 数据可视化
matplotlib>=3.7.0

# 注意：tkinter 是 Python 标准库的一部分，不需要单独安装